# openapi.yaml
openapi: 3.1.0
info:
  title: OKR Tracking System API
  description: |
    ## OKR Tracking System — Garant Bank

    REST API for managing **Objectives & Key Results (OKR)**, organizational structure, user management, and multi-source performance evaluations.

    ### Key Features
    - **OKR Management** — Create and track Objectives with weighted Key Results
    - **Multi-Source Evaluations** — Director (star rating), HR (letter grade), and Business Block assessments
    - **Score Calculation** — Automatic weighted scoring: 60% OKR + 20% Director + 20% HR
    - **Organizational Hierarchy** — Divisions → Departments → Objectives → Key Results
    - **Role-Based Access Control** — 6 roles with fine-grained permissions
    - **Excel Export** — Export all OKR data to Excel spreadsheets

    ### Authentication
    All endpoints (except login and export) require a **JWT Bearer token**.
    1. Call `POST /api/auth/login` with your credentials
    2. Copy the `token` from the response
    3. Click the **Authorize** button above and enter the token

    ### Roles
    | Role | Description |
    |------|-------------|
    | `ADMIN` | Full system access, user management |
    | `DIRECTOR` | Division oversight, star-based evaluations |
    | `DEPARTMENT_LEADER` | Department OKR management |
    | `HR` | Letter-grade evaluations (A/B/C/D) |
    | `BUSINESS_BLOCK` | Business block star-based evaluations |
    | `EMPLOYEE` | View own profile and assigned departments |
  version: 1.0.0
...
servers:
...
security:
- Bearer Authentication: []
...
paths:
...
  /api/score-levels:
    get:
      tags:
      - Score Levels
      summary: Get all score levels
      description: "Returns all configured score levels ordered by display order.\
        \ Default levels: Below (3.0), Meets (4.25), Good (4.5), Very Good (4.75),\
        \ Exceptional (5.0)."
      operationId: getAllScoreLevels
      responses:
        "200":
          description: List of score levels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScoreLevelDTO"
    put:
      tags:
      - Score Levels
      summary: Update score levels
      description: "Replace all score levels with the provided list. Each level must\
        \ have a unique name, score value, color, and display order."
      operationId: updateScoreLevels
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/ScoreLevelDTO"
        required: true
      responses:
        "200":
          description: Score levels updated — returns new configuration
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScoreLevelDTO"
        "400":
          description: Invalid configuration
...
components:
  schemas:
...
    ScoreLevelDTO:
      type: object
      description: Score level configuration entry
      properties:
        id:
          type: string
          description: Score level ID
          example: sl-001
        name:
          type: string
          description: Level name
          enum:
          - Below
          - Meets
          - Good
          - Very Good
          - Exceptional
          example: Good
        scoreValue:
          type: number
          format: double
          description: Numeric score value for this level
          example: 4.5
        color:
          type: string
          description: Hex color code for UI display
          example: "#4CAF50"
        displayOrder:
          type: integer
          format: int32
          description: Display order (lower = displayed first)
          example: 3
...
  securitySchemes:
    Bearer Authentication:
      type: http
      description: Enter the JWT token obtained from `POST /api/auth/login`
      name: Bearer Authentication
      scheme: bearer
      bearerFormat: JWT